Added wimmelpicture

+++ pmw_mouse_sensor/waldo.h
@@ -3,7 +3,7 @@
 
 #include <M5Stack.h>
 
-void updateWaldo(int16_t x, int16_t y);
+void updateWaldo(int32_t x, int32_t y);
 
 void initWaldo();
